changed arrowInputController to keyboardInputController
